+++ src/card_group/deck.rs
@@ -1,4 +1,4 @@
-use super::card_group::CardGroup;
+use super::CardGroup;
 use std::fmt;
 
 pub struct Deck {
@@ -9,9 +9,13 @@ impl Deck {
     pub fn new() -> Deck {
         let mut card_group = CardGroup::all_cards();
         card_group.shuffle();
-        return Deck {
-            card_group: card_group,
-        };
+        Deck { card_group }
+    }
+}
+
+impl Default for Deck {
+    fn default() -> Self {
+        Self::new()
     }
 }
